Add files via upload
The Home Care Services for Senior’s project is a web-based application developed on the .Net framework, specifically designed to cater to the distinctive healthcare requirements of senior citizens. Focused on delivering comprehensive and personalized care, this platform empowers seniors to receive a variety of home-based healthcare services, ensuring comfort, convenience, and an enhanced quality of life and do the payment online.